+++ src/qsys/Camera.cpp
@@ -256,12 +256,12 @@ namespace {
       if (m_nMode==VSE_ADD) {
         // remove
         pCam->visRemove(m_nElemID);
-        MB_DPRINTLN("VSE.undo> VSE_ADD remove(%d)", m_nElemID);
+        MB_DPRINTLN("VSE.undo> VSE_ADD remove(%d)", (int)m_nElemID);
       }
       else {
         // add
         pCam->visAppend(m_nElemID, m_value.bVis, m_value.bObj);
-        MB_DPRINTLN("VSE.undo> VSE_REMOVE append(%d, %d, %d)", m_nElemID, m_value.bVis, m_value.bObj);
+        MB_DPRINTLN("VSE.undo> VSE_REMOVE append(%d, %d, %d)", (int)m_nElemID, m_value.bVis, m_value.bObj);
       }
       return true;
     }
@@ -273,12 +273,12 @@ namespace {
       if (m_nMode==VSE_ADD) {
         // add
         pCam->visAppend(m_nElemID, m_value.bVis, m_value.bObj);
-        MB_DPRINTLN("VSE.redo> VSE_ADD append(%d, %d, %d)", m_nElemID, m_value.bVis, m_value.bObj);
+        MB_DPRINTLN("VSE.redo> VSE_ADD append(%d, %d, %d)", (int)m_nElemID, m_value.bVis, m_value.bObj);
       }
       else {
         // remove
         pCam->visRemove(m_nElemID);
-        MB_DPRINTLN("VSE.redo> VSE_REMOVE remove(%d)", m_nElemID);
+        MB_DPRINTLN("VSE.redo> VSE_REMOVE remove(%d)", (int)m_nElemID);
       }
       return true;
     }
