+++ src/qsys/View.hpp
@@ -526,7 +526,7 @@ namespace qsys {
 
     void checkAndUpdate() {
         if (m_bUpdateRequired) {
-            MB_DPRINTLN("View::checkAndUpdate> view %d update %d", m_uid, m_bUpdateRequired);
+            MB_DPRINTLN("View::checkAndUpdate> view %d update %d", (int)m_uid, m_bUpdateRequired);
             drawScene();
         }
         clearUpdateFlag();
